Fix Prometheus metrics cleanup on Celery worker exit
Celery workers can be started and stopped depending on the load.  When
they're stopped, Prometheus metric files must be cleaned up.  This is
done by calling the mark_process_dead from the prometheus_client library
in a similar way to how it's done for exiting gunicorn processes.

Note that the call deletes some of the .db files created by the process,
but not all of them.  Some of them must be kept to properly track global
(not process-specific) metrics, e.g. counters.