Fix race condition in Prometheus metrics directory clean up
Containers defined in docker-compose.yml can use Prometheus metrics. As
it's a multiprocess setup, the metrics are synchronized using files on
disk between different processes.  Until now, the main app container
with the Django application, cleaned up this directory (recursively)
from any files which might have remained from previous runs.

Unfortunately, deleting all the files means that this might delete
Prometheus metrics collected in processes running in other containers.
This causes them to crash trying to access non-existent files on disk.

Deleting the Prometheus files is only allowed when the whole application
(i.e. all its containers) are stopped -- usually when docker-compose
down is run.  In other scenarios process crashes can occur or incorrect
metrics can be observed.

The only files which can be deleted are the ones holding gauge metrics
with multiprocess_mode set to live*.  These are also deleted when
multiprocess.mark_process_dead is called.

This commit fixes the race by only deleting the files with gauge live*
metrics on container start up.  Each container only cleans up its own
files.

At the same time, the metrics files are now stored in a named volume
defined in docker-compose.yml.  This way, docker-compose will guarantee
that the directory is empty on first run.  When `docker-compose down -v`
is run, the volume will be deleted with all metrics, but this will be
done after all containers stop.

Note that this also means that most of the metrics files will remain
in the volume if `docker-compose down` will be called without the `-v`
switch.  This is OK however -- after restart the newly restarted
processes will start appending to the same metric files.  In other words
-- processes will continue counting from the last run.  Note that in
terms of metrics and the Prometheus server perspective this is
equivalent to resetting back to the initial state.

The only drawback of this solution is that the number of metric files
will keep growing.  This will slow down scrapping.  This is however an
independet issue and will happen anyway.  Even if the whole application
is not restarted, individual containers will start and stop new
processes (e.g. gunicorn or celery workers) and each of them will add
extra files.