Make celery-worker container exit if celery crashes
The celery-worker container runs workers in background processes.  The
process that runs in foreground is a tail command on its logfiles.  If
the celery workers crash (or fail to start up), the container will keep
running the tail command instead of exiting.

This commit fixes the problem by running the tail command in background
and running a loop with periodic celery status checks.  This duplicates
the healthcheck defined in the docker-compose file, but seems to still
be a reasonable solution as the `celery multi` command doesn't support
running in foreground.

Note that Celery can crash if the main `app` container with the Django
application restars.  Just before startup, the entrypoint.sh script
deletes the contents of the directory with the Prometheus metrics.  If
the celery-worker container starts up first, this can delete its
Prometheus files causing a crash.  This fix will ensure the
celery-worker will be properly restarted if that happens.

This problem could be solved by simply specifying a startup order using
the `depends_on` keyword.  However, it would only help in the case where
all containers were starting up together (i.e. after docker-compose up).
If the app container was restarted alone, the same problem would occur.